Move C# generator to under the libbun.encode.release
