+++ java/src/libbun/util/LibBunSystem.java
@@ -568,7 +568,7 @@ public class LibBunSystem {
 		ClassMap.put("bun", libbun.encode.playground.BunGenerator.class);
 		ClassMap.put("c",   libbun.encode.playground.CGenerator.class);
 		ClassMap.put("cl",  libbun.encode.playground.CommonLispGenerator.class);
-		ClassMap.put("cs",  libbun.encode.playground.CSharpGenerator.class);
+		ClassMap.put("cs",  libbun.encode.release.CSharpGenerator.class);
 		ClassMap.put("erl", libbun.encode.erlang.ErlangGenerator.class);
 
 		ClassMap.put("hs",  libbun.encode.haskell.HaskellSourceGenerator.class);
